requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
opencv-python==4.8.1.78
Pillow==10.0.1
google-generativeai==0.3.2
boto3==1.29.7
pydub==0.25.1